components/lambda/requirements.txt
google-cloud-storage
# aws_lambda_powertools
google-cloud-container